+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - show release notes when npm and gh release match
 - add control modal for filter to reset filter life
 - show bool characteristic descriptions in info modals
+- show colour temp values in mired and kelvin
 
 ### Homebridge Dependencies